+++ tests/src/IntegrationTests/RecollectionDecoratorSmallSetTest.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace Rekalogika\Collections\Tests\IntegrationTests;
 
 use Doctrine\ORM\PersistentCollection;
-use Rekalogika\Collections\Tests\App\BasicRepository\CountryMinimalRepository;
 use Rekalogika\Collections\Tests\App\Entity\Citizen;
 use Rekalogika\Collections\Tests\App\Entity\Country;
+use Rekalogika\Collections\Tests\App\MinimalRepository\CountryMinimalRepository;
 use Rekalogika\Collections\Tests\IntegrationTests\Base\BaseRecollectionTestCase;
 use Rekalogika\Collections\Tests\IntegrationTests\Trait\RecollectionTestsTrait;
 use Rekalogika\Contracts\Collections\MinimalRepository;